Agrego servicios y controllers de archivo y materia, ajustes en sus dtos, repositorios y model

+++ src/main/java/com/challenge/aranguriAppsBackend/dto/ArchivoResponse.java
@@ -1,3 +1,3 @@
 package com.challenge.aranguriAppsBackend.dto;
 
-public record ArchivoResponse(Long id, String nombre, String tipo, String url) { }
+public record ArchivoResponse(int id, String nombre, String tipo, String url) { }
